Update training script and model processing for improved validation
- Modified `train.py` to increase `num_sanity_val_steps` from 0 to 2, enhancing the validation process during training.
- Updated waveform processing in `LatentDiffusion` class to ensure the output is converted to float before converting to numpy, improving compatibility with downstream tasks.

+++ models/flowsep/model.py
@@ -906,7 +906,7 @@ class LatentDiffusion(DDPM):
             mel = mel.squeeze(1)
         mel = mel.permute(0, 2, 1)
         waveform = self.first_stage_model.vocoder(mel)
-        waveform = waveform.cpu().detach().numpy()
+        waveform = waveform.cpu().detach().float().numpy()
         if save:
             self.save_waveform(waveform, savepath, name)
         return waveform
